# adex-rest/src/main/java/com/adex/demo/rest/api/CustomerRequestApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech Do not edit the class manually.
 */
package com.adex.demo.rest.api;

import com.adex.demo.rest.models.CustomerRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import javax.annotation.processing.Generated;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@Generated(
    value = "org.openapitools.codegen.languages.SpringCodegen",
    comments =
        "Generator version: 7.5.0, using custom templates: /src/main/resources/openapitemplates")
@Tag(name = "CustomerRequest", description = "Sample Request.")
public interface CustomerRequestApi {

  /**
   * POST /api/v1/customer Sample request.
   *
   * @param customerRequest (required)
   * @return Accepted. (status code 202)
   */
  @Operation(
      operationId = "processCustomerRequest",
      description = "Sample request.",
      tags = {"CustomerRequest"},
      responses = {@ApiResponse(responseCode = "202", description = "Accepted.")})
  @RequestMapping(
      method = RequestMethod.POST,
      value = "/api/v1/customer",
      consumes = "application/json")
  ResponseEntity<Void> processCustomerRequest(
      @Parameter(name = "CustomerRequest", description = "", required = true) @RequestBody
          CustomerRequest customerRequest);
}